Add files via upload
There is a definite bug in this code, but I've been unable to find it. Basically, it brings in a small number of redundant attributes and a small number of redundant citations, but I can't begin to figure out what is causing that. I have a dummy line in Rep (println("Break")) that identifies one such "issue" point, but I can't seem to figure out why it only activates once when it should activate no less than six times. Also, I think the read in of the attributes and citations of the State place cuts off after one or a few, but I haven't identified the problem.